+++ R/jamenrich-igraph.r
@@ -1969,3 +1969,157 @@ reorderIgraphNodes <- function
    V(g)[iMatch]$y <- newDF$y;
    return(g);
 }
+
+#' Remove igraph singlet nodes
+#'
+#' Remove igraph singlet nodes
+#'
+#' This function is a lightweight method to remove igraph nodes with
+#' no connections. In fact, the `min_degree` can be used to require
+#' a minimum number of connections, but the intended use is to remove
+#' the singlet nodes that have no connections.
+#'
+#' @family jam igraph functions
+#'
+#' @param g igraph object
+#' @param min_degree numeric threshold with the minimum number of
+#'    connections, also known as the "degree", required for each node.
+#' @param ... additional arguments are ignored.
+#'
+#' @export
+removeIgraphSinglets <- function
+(g,
+ min_degree=1,
+ ...)
+{
+   keep_nodes <- (degree(g) >= min_degree);
+   g_new <- igraph::subgraph(g,
+      which(keep_nodes));
+   return(g_new);
+}
+
+#' Spread igraph node labels by angle from node center
+#'
+#' Spread igraph node labels by angle from node center
+#'
+#' This function uses the igraph vertex attribute
+#' `"label.degree"`, which describes the angular offset for
+#' each vertex label. (The `"label.degree"` values are
+#' represented as radians, not degrees, starting at 0 for
+#' right, and proceeding clockwise starting from the right,
+#' down, left, top, right.)
+#'
+#' This function requires a network layout, which must be fixed
+#' in order for the vertex labels to be properly oriented.
+#' Labels are oriented opposite the most dominant angular mean
+#' of edges from each network node. Typically the side of a node
+#' with the fewest edges has the most space to place a label.
+#' No further checks are performed for overlapping labels.
+#'
+#' Note that this function does not modify the other important
+#' attribute `"label.distance"`, which is required in order to
+#' see any of these changes. If `label_distance=0` then the label
+#' will be centered on top of each igraph node.
+#'
+#' @jam igraph functions
+#'
+#' @param g igraph object
+#' @param layout numeric matrix representing the x and y
+#'    coordinates of each node in `g`, in the same order as `V(g)`.
+#'    When `layout` is not supplied, nodes are checked for
+#'    attributes `c("x", "y")` which define a fixed internal
+#'    layout. If that is not supplied, then `layout_with_qfr()`
+#'    is called along with the `repulse` argument. Subsequent
+#'    coordinates are stored in `V(g)$x` and `V(g)$y` when
+#'    argument `update_g_coords=TRUE`.
+#' @param y_bias numeric value indicating the tendency to spread
+#'    labels on the y-axis rather than symmetrically around each node.
+#'    This argument elongates the circle surrounding a node into
+#'    an ellipse with this ratio.
+#' @param update_g_coords logical indicating whether the layout
+#'    coordinates will be stored in `V(g)$x` and `V(g)$y`.
+#' @param do_reorder logical indicating whether to call
+#'    `reorderIgraphNodes()` which re-distributes equivalent nodes
+#'    based upon the node color(s). A node is "equivalent" to another
+#'    node if both nodes have identical edges.
+#' @param sortAttributes,nodeSortBy arguments passed to
+#'    `reorderIgraphNodes()` when `do_reorder=TRUE`.
+#' @param repulse argument passed to `layout_with_qfr()` only
+#'    when `layout` is not supplied, and the layout is not stored
+#'    in `c(V(g)$x, V(g)$y)`.
+#' @param label_min_dist numeric value used to ensure all labels are
+#'    at least some distance from the center. These units are defined
+#'    by igraph, and are roughly in units of one line height of text.
+#' @param ... additional arguments are passed to `layout_with_qfr()`
+#'    when needed.
+#'
+#' @export
+spread_igraph_labels <- function
+(g,
+ layout=NULL,
+ y_bias=3,
+ update_g_coords=TRUE,
+ do_reorder=TRUE,
+ sortAttributes=c("pie.color", "coloredrect.color", "color"),
+ nodeSortBy=c("x","y"),
+ repulse=3.5,
+ label_min_dist=0.5,
+ verbose=FALSE,
+ ...)
+{
+   ##
+   if (length(layout) == 0) {
+      if (all(c("x", "y") %in% list.vertex.attributes(g))) {
+         layout <- cbind(x=V(g)$x, y=V(g)$y);
+      } else {
+         layout <- layout_with_qfr(g,
+            repulse=repulse,
+            ...);
+      }
+   } else if (is.function(layout)) {
+      layout <- layout(g);
+   }
+   if (length(rownames(layout)) == 0) {
+      rownames(layout) <- V(g)$name;
+   }
+   if (do_reorder) {
+      if (verbose) {
+         printDebug("spread_igraph_labels(): ",
+            "Calling multienrichjam::reorderIgraphNodes()");
+         printDebug("spread_igraph_labels(): ",
+            "head(layout) before:");
+         print(head(layout));
+      }
+      g <- multienrichjam::reorderIgraphNodes(g,
+         layout=layout,
+         nodeSortBy=c("x","y"),
+         sortAttributes=sortAttributes);
+      layout <- cbind(V(g)$x, V(g)$y);
+      if (verbose) {
+         printDebug("spread_igraph_labels(): ",
+            "head(layout) after:");
+         print(head(layout));
+      }
+   }
+   g_angle <- nameVector(sapply(seq_len(vcount(g)), function(i){
+      xy1 <- layout[i,,drop=FALSE];
+      xy2 <- layout[as.numeric(ego(g, nodes=i, mindist=1)[[1]]),,drop=FALSE];
+      if (length(xy2) == 0) {
+         xy2 <- matrix(ncol=2, c(0,0));
+      }
+      xymean <- colMeans(xy1[rep(1, nrow(xy2)),,drop=FALSE] - xy2);
+      -(xyAngle(xymean[1], xymean[2]*y_bias, directed=TRUE) + 0) %% 360
+   }), V(g)$name);
+   if (update_g_coords) {
+      V(g)$x <- layout[,1];
+      V(g)$y <- layout[,2];
+   }
+   V(g)$label.degree <- deg2rad(g_angle);
+   if (!"label.dist" %in% list.vertex.attributes(g)) {
+      V(g)$label.dist <- label_min_dist;
+   } else {
+      V(g)$label.dist <- pmax(V(g)$label.dist, label_min_dist);
+   }
+   g;
+}
+